requirement.txt
anyio==4.5.2
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.13.3
bleach==6.1.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
Click==7.0
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.12
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.19
djangorestframework==3.15.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
fonttools==4.55.8
fsspec==2025.2.0
future==0.18.2
GDAL==3.3.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httpx==0.28.1
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.1
monotonic==1.5
more-itertools==4.2.0
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
notebook==7.3.2
notebook-shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.50
protobuf==3.6.1
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pygments==2.19.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-json-logger==3.2.1
pytz==2025.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.2.1
referencing==0.35.1
reportlab==3.5.34
requests==2.32.3
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.3
simple-lama-inpainting==0.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.6.2
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.2.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ultralytics==8.3.70
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xgboost==2.1.4
xkit==0.0.0
zipp==3.20.2
